+++ lib/taro/types/shared/deprecation.rb
@@ -1,3 +1,7 @@
 module Taro::Types::Shared::Deprecation
-  attr_accessor :deprecated
+  attr_reader :deprecated
+
+  def deprecated=(value)
+    @deprecated = value ? true : nil
+  end
 end